# Optimize escape_single_quoted_newlines